# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_giovanysievert_ask-anything_internal_httputil.ErrorBody:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorBody'
    type: object
  internal_chat.Conversation:
    properties:
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  internal_chat.CreateConversationRequest:
    properties:
      title:
        example: React Native study
        maxLength: 255
        type: string
    type: object
  internal_chat.Message:
    properties:
      content:
        type: string
      conversation_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      role:
        type: string
    type: object
  internal_chat.MessageRequest:
    properties:
      content:
        example: How do I optimize a slow FlatList?
        minLength: 1
        type: string
    required:
    - content
    type: object
  internal_document.CreateRequest:
    properties:
      content:
        example: FlatList renders large lists efficiently. Use getItemLayout, keyExtractor,
          windowSize...
        minLength: 1
        type: string
      title:
        example: React Native FlatList
        maxLength: 255
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
  internal_document.Document:
    properties:
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
  description: AI-powered chat API. Ingest documents, then hold multi-turn conversations
    whose answers are grounded in the ingested material (RAG) and streamed back over
    Server-Sent Events.
  title: ask-anything API
  version: "1.0"
paths:
  /conversations:
    get:
      description: Returns conversations ordered by most recently updated.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_chat.Conversation'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: List conversations
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Starts a new chat conversation.
      parameters:
      - description: Optional title
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_chat.CreateConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_chat.Conversation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: Create a conversation
      tags:
      - chat
  /conversations/{id}/messages:
    get:
      description: Returns the full message history of a conversation, oldest first.
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_chat.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: List messages in a conversation
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: 'Appends the user message, retrieves grounding chunks (RAG), and
        streams the assistant reply as Server-Sent Events. Emits `event: delta` frames
        with a JSON `{"text":"..."}` payload, a terminal `event: done` frame carrying
        the persisted assistant message, and `event: error` on failure.'
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: string
      - description: User message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_chat.MessageRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: Send a message and stream the reply (SSE)
      tags:
      - chat
  /documents:
    get:
      parameters:
      - description: Max documents to return (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: Number of documents to skip
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_document.Document'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: List ingested documents
      tags:
      - documents
    post:
      consumes:
      - application/json
      description: Chunks the content, embeds each chunk locally via Ollama, and stores
        the vectors for retrieval.
      parameters:
      - description: Title and content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_document.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_document.Document'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_giovanysievert_ask-anything_internal_httputil.ErrorResponse'
      summary: Ingest a document
      tags:
      - documents
swagger: "2.0"